requirements.txt
streamlit
pandas
firebase-admin
openai
python-dotenv